A rogue uint that was missed.
git-svn-id: http://svn.atnf.csiro.au/duchamp/trunk@718 450e7745-72da-457f-8a74-b07dc0cda934

+++ src/Cubes/existingDetections.cc
@@ -105,7 +105,7 @@ namespace duchamp
       if(temp.substr(0,11)!="Detection #") duchampError("existingDetections","Format of Log file is wrong!");
       Detection obj;
       while(getline(logfile,temp), temp.size()>0){
-	  for(uint i=0;i<temp.size();i++)
+	  for(size_t i=0;i<temp.size();i++)
 	    if(temp[i]=='-' || temp[i]==',') temp[i] = ' ';
 	  std::stringstream ss;
 	  ss.str(temp);
